renaming file so that it doesn't collide with setup in the python toolbox
