+++ auto_setup.py
